Fix the storymap grid.
Before, some stories were being inserted into the wrong part
of the grid. This was due to an incorrect assumption that the
pos property of the epic and sprint objects accurately reflected
the actual positions of the sprints and epics on the rendered
grid.

Now, the sprint and epic data containers are arrays, and
the positions of the epics and sprints are determined in the
renderMap function.