Fix issue with story modal loading.
Before, the loadStory function grabbed a story from the
storiesList data container and modified it. Then, it
passed the modified story to the template as a context.
If a user attempted to load the same story again, an error
would be raised, because the loadStory function accesses
a modified version of the story from the previous call.

Now, the properties of the story are copied to a new context
object, and that new context object is passed to the template
instead.